+++ src/readbytedata.cc
@@ -4,22 +4,20 @@
 #include "./i2c-dev.h"
 #include "./readbytedata.h"
 
-static int ReadByteData(int fd, int cmd) {
+static __s32 ReadByteData(int fd, __u8 cmd) {
   return i2c_smbus_read_byte_data(fd, cmd);
 }
 
 class ReadByteDataWorker : public NanAsyncWorker {
 public:
-  ReadByteDataWorker(NanCallback *callback, int fd, int cmd)
+  ReadByteDataWorker(NanCallback *callback, int fd, __u8 cmd)
     : NanAsyncWorker(callback), fd(fd), cmd(cmd) {}
   ~ReadByteDataWorker() {}
 
   void Execute() {
-    int ret = ReadByteData(fd, cmd);
-    if (ret == -1) {
+    byte = ReadByteData(fd, cmd);
+    if (byte == -1) {
       SetErrorMessage(strerror(errno));
-    } else {
-      val = ret;
     }
   }
 
@@ -28,7 +26,7 @@ public:
 
     v8::Local<v8::Value> argv[] = {
       NanNull(),
-      NanNew<v8::Integer>(val)
+      NanNew<v8::Integer>(byte)
     };
 
     callback->Call(2, argv);
@@ -36,8 +34,8 @@ public:
 
 private:
   int fd;
-  int cmd;
-  int val;
+  __u8 cmd;
+  __s32 byte;
 };
 
 NAN_METHOD(ReadByteDataAsync) {
@@ -48,7 +46,7 @@ NAN_METHOD(ReadByteDataAsync) {
   }
 
   int fd = args[0]->Int32Value();
-  int cmd = args[1]->Int32Value();
+  __u8 cmd = args[1]->Int32Value();
   NanCallback *callback = new NanCallback(args[2].As<v8::Function>());
 
   NanAsyncQueueWorker(new ReadByteDataWorker(callback, fd, cmd));
@@ -63,13 +61,13 @@ NAN_METHOD(ReadByteDataSync) {
   }
 
   int fd = args[0]->Int32Value();
-  int cmd = args[1]->Int32Value();
+  __u8 cmd = args[1]->Int32Value();
 
-  int ret = ReadByteData(fd, cmd);
-  if (ret == -1) {
+  __s32 byte = ReadByteData(fd, cmd);
+  if (byte == -1) {
     return NanThrowError(strerror(errno), errno);
   }
 
-  NanReturnValue(NanNew<v8::Integer>(ret));
+  NanReturnValue(NanNew<v8::Integer>(byte));
 }